chore(wms): move shared movement enums to domain package
