+++ app/wms/scan/services/pick_handler.py
@@ -1,7 +1,7 @@
 # app/wms/scan/services/pick_handler.py
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.models.enums import MovementType
+from app.wms.shared.enums import MovementType
 from app.wms.stock.services.stock_service import StockService
