+++ app/wms/stock/services/stock_ship_service.py
@@ -8,7 +8,7 @@ from sqlalchemy import text
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.core.problem import raise_problem
-from app.models.enums import MovementType
+from app.wms.shared.enums import MovementType
 from app.wms.outbound.services.invariant_guard_outbound import enforce_outbound_invariant_guard
 from app.wms.shared.services.lot_code_contract import fetch_item_expiry_policy_map
