+++ app/wms/warehouses/models/warehouse.py
@@ -8,6 +8,7 @@ from sqlalchemy import Boolean, Integer, String, text
 from sqlalchemy.orm import Mapped, mapped_column, relationship, validates
 
 from app.db.base import Base
+from app.tms.providers.models.warehouse_shipping_provider import WarehouseShippingProvider  # noqa: F401
 
 if TYPE_CHECKING:
     from app.tms.providers.models.warehouse_shipping_provider import WarehouseShippingProvider
