+++ tests/quick/test_inbound_reclassify_v2.py
@@ -4,7 +4,7 @@ import pytest
 from sqlalchemy import text
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.models.enums import MovementType
+from app.wms.shared.enums import MovementType
 from app.wms.stock.services.stock_service import StockService
 
 UTC = timezone.utc
