+++ tests/unit/test_fefo_query_v2.py
@@ -6,7 +6,7 @@ from datetime import date, datetime, timedelta, timezone
 import pytest
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.models.enums import MovementType
+from app.wms.shared.enums import MovementType
 from app.wms.shared.services.expiry_analytics_allocator import ExpiryAnalyticsAllocator
 from app.wms.stock.services.stock_service import StockService
 from tests.utils.ensure_minimal import ensure_item
